Optimized map rendering and other enhancements
- Map rendering per frame went from 100ms to very short time. This will
allow for other activities.
- Added frame rate limiting
- CPU utilization went from 100% down to 2% with map rendering
optimizations
- Added start of Sprite class
- Fixed some warnings in 3rd party code

NOTE: Only Xcode project updated, codelite has not been updated

+++ tmxparser/TmxLayer.cpp
@@ -203,7 +203,7 @@ namespace Tmx
 			// Use the utility class for decompressing (which uses zlib)
 			out = (unsigned *)Util::DecompressGZIP(
 				text.c_str(),
-				text.size(),
+				static_cast<int>(text.size()),
 				width * height * 4);
 		}
 		else
